requirements.txt
requests==2.18.4
requests-toolbelt==0.7.0
twilio==6.9.0
urllib3==1.22
firebase_admin==2.8.0
python-dotenv==0.7.1
Flask==0.12.2
gunicorn==19.7.1
fuzzywuzzy==0.16.0